enforce mrn data type to avoid join issue
When reading the JSON files (for bronze table), mrn data is read as a string. This will cause a join issue in the later part as mrn is read as integer in pii table. 

Note: I realise that all other PRs are asked (@SireInsectus ) to raise the issue in "Curriculum-Dev" workspace. But I'm not sure what that means so raising a PR here regardless.